+++ packages/reactabular-utils/src/columns-are-equal.js
@@ -1,0 +1,12 @@
+import isFunction from 'lodash/isFunction';
+import diff from 'deep-diff';
+
+function columnsAreEqual(oldColumns, newColumns) {
+  return (diff(
+    oldColumns, newColumns
+  ) || []).filter(
+    e => !isFunction(e.lhs)
+  ).length === 0;
+}
+
+export default columnsAreEqual;