+++ packages/reactabular-utils/src/rows-are-equal.js
@@ -1,0 +1,7 @@
+import isEqual from 'lodash/isEqual';
+
+function rowsAreEqual(oldRows, newRows) {
+  return isEqual(oldRows, newRows);
+}
+
+export default rowsAreEqual;